A monster change. Whilst writing the scalar regtest I got fed up with hunting through the syswrap files and headers for the code. There was some grouping by functionality. With this change all syscalls are ordered by syscall number. In the main syswrap I've also put the syscall number and the associated function copy/pasted from the manpage (if such exists; sometimes the real interface differs).
In doing so I unearthed at least one duplicate (in an #if 0 block),
and added numerous syscalls. I changed the argument names to match
thos found in the manpage. That's probably more useful than either
the wroong name copied from Linux or the name in syscalls.master.

The scalar testcase has been greatly expanded, not quite yet done
amd there are several todos. On the way I fixed several crashes
in Valgrind due to dereferencing invalid pointers. Testing
"if (pter != NULL)" is insufficient, it needs to be done with the
"is_safe_to_deref" function.

On the downside this will zap a lot of the git history. It won't
be the last time.