s390x regtest: Refactor mul test cases for Clang support
The mul test case crashes when compiled with clang because of register
clashes in some inline assemblies.  Obviously clang does not treat
register clobbers as "early clobbers", but sometimes uses the same
registers for inputs as well.

Rewrite the inline assemblies in mul.h to avoid this issue.  Also, reduce
the code duplication in the process.  Adjust mul_GE.c accordingly.  Keep
the expected output for all mul test cases the same as before.